suppress trivial warnings

+++ core/src/Physics/PHHapticEngine.cpp
@@ -65,7 +65,7 @@ namespace Spr {
 		stribeckVelocity = (shape[0]->GetMaterial().stribeckVelocity + shape[1]->GetMaterial().stribeckVelocity) * 0.5;
 		stribeckmu = (shape[0]->GetMaterial().stribeckmu + shape[1]->GetMaterial().stribeckmu) * 0.5;
 
-		for (int i = 0; i < shape[0]->GetMaterial().timeVaryFrictionAs.size(); i++) {
+		for (int i = 0; i < (int) shape[0]->GetMaterial().timeVaryFrictionAs.size(); i++) {
 			mus[i] = (shape[0]->GetMaterial().mus[i] + shape[1]->GetMaterial().mus[i]) * 0.5;
 			mu0s[i] = (shape[0]->GetMaterial().mu0s[i] + shape[1]->GetMaterial().mu0s[i]) * 0.5;
 			timeVaryFrictionAs[i] = (shape[0]->GetMaterial().timeVaryFrictionAs[i] + shape[1]->GetMaterial().timeVaryFrictionAs[i]) * 0.5;
